Remove old-style Python-C++ converters
And merge the two files now that there are no old ones.
